Introduce the concept of capabilities
Capabilities are a new first-class concept in Gradle, living along variants. They can be used to model
components which "provide the same capability". For example, two different implementations of the same
API can provide the same capability, but it's illegal to have both on the dependency graph: the user
must choose between one of the two.

This commit introduces the ability to declare capabilities at the variant level, for now only using
component metadata rules. Capabilities participate in conflict resolution: as soon as we find that two
modules in the graph provide the same capability, we fail. There's currently no way to prefer one over
the other (this is going to be implemented in a subsequent commit). This means that if such a conflict
occurs, the user must select the right module appropriately using component selection rules or module
metadata rules.